logical-monitor: Make scale a float
To support fractional scaling, the logical monitor scale must be stored
as a float. No other functional changes is part of this commit.

https://bugzilla.gnome.org/show_bug.cgi?id=765011

+++ src/backends/meta-logical-monitor.h
@@ -40,7 +40,7 @@ struct _MetaLogicalMonitor
   gboolean is_primary;
   gboolean is_presentation; /* XXX: not yet used */
   gboolean in_fullscreen;
-  int scale;
+  float scale;
   MetaMonitorTransform transform;
 
   /* The primary or first output for this monitor, 0 if we can't figure out.
@@ -68,7 +68,7 @@ MetaLogicalMonitor * meta_logical_monitor_new (MetaMonitorManager       *monitor
 MetaLogicalMonitor * meta_logical_monitor_new_derived (MetaMonitorManager *monitor_manager,
                                                        MetaMonitor        *monitor,
                                                        MetaRectangle      *layout,
-                                                       int                 scale,
+                                                       float               scale,
                                                        int                 monitor_number);
 
 void meta_logical_monitor_add_monitor (MetaLogicalMonitor *logical_monitor,
@@ -78,7 +78,7 @@ gboolean meta_logical_monitor_is_primary (MetaLogicalMonitor *logical_monitor);
 
 void meta_logical_monitor_make_primary (MetaLogicalMonitor *logical_monitor);
 
-int meta_logical_monitor_get_scale (MetaLogicalMonitor *logical_monitor);
+float meta_logical_monitor_get_scale (MetaLogicalMonitor *logical_monitor);
 
 GList * meta_logical_monitor_get_monitors (MetaLogicalMonitor *logical_monitor);
